+++ tslearn/bases/__init__.py
@@ -1,0 +1,3 @@
+from .bases import TimeSeriesBaseEstimator, BaseModelPackage
+
+__all__ = ["TimeSeriesBaseEstimator", "BaseModelPackage"]
