+++ tslearn/datasets/ucr_uea.py
@@ -1,83 +1,11 @@
-"""
-The :mod:`tslearn.datasets` module provides simplified access to standard time
-series datasets.
-"""
-
-import numpy
-import zipfile
-import tempfile
-import shutil
 import os
-import csv
 import warnings
+import csv
+import shutil
 from urllib.request import urlretrieve
 
 from tslearn.utils import _load_arff_uea, _load_txt_uea
-
-__author__ = 'Romain Tavenard romain.tavenard[at]univ-rennes2.fr'
-
-
-def extract_from_zip_url(url, target_dir=None, verbose=False):
-    """Download a zip file from its URL and unzip it.
-
-    A `RuntimeWarning` is printed on failure.
-
-    Parameters
-    ----------
-    url : string
-        URL from which to download.
-    target_dir : str or None (default: None)
-        Directory to be used to extract unzipped downloaded files.
-    verbose : bool (default: False)
-        Whether to print information about the process (cached files used, ...)
-
-    Returns
-    -------
-    str or None
-        Directory in which the zip file has been extracted if the process was
-        successful, None otherwise
-    """
-    fname = os.path.basename(url)
-    tmpdir = tempfile.mkdtemp()
-    local_zip_fname = os.path.join(tmpdir, fname)
-    urlretrieve(url, local_zip_fname)
-    os.makedirs(target_dir, exist_ok=True)
-    try:
-        with zipfile.ZipFile(local_zip_fname, "r") as f:
-            f.extractall(path=target_dir)
-        if verbose:
-            print("Successfully extracted file %s to path %s" %
-                  (local_zip_fname, target_dir))
-        return target_dir
-    except zipfile.BadZipFile:
-        warnings.warn("Corrupted or missing zip file encountered, aborting",
-                      category=RuntimeWarning)
-        return None
-    finally:
-        shutil.rmtree(tmpdir, ignore_errors=True)
-
-
-def in_file_string_replace(filename, old_string, new_string):
-    """String replacement within a text file. It is used to fix typos in
-    downloaded csv file.
-
-    The code was modified from "https://stackoverflow.com/questions/4128144/"
-
-    Parameters
-    ----------
-    filename : str
-        Path to the file where strings should be replaced
-    old_string : str
-        The string to be replaced in the file.
-    new_string : str
-        The new string that will replace old_string
-    """
-    with open(filename) as f:
-        s = f.read()
-
-    with open(filename, 'w') as f:
-        s = s.replace(old_string, new_string)
-        f.write(s)
+from .datasets import in_file_string_replace, extract_from_zip_url
 
 
 class UCR_UEA_datasets:
@@ -421,84 +349,3 @@ class UCR_UEA_datasets:
                               category=RuntimeWarning, stacklevel=2)
 
 
-class CachedDatasets:
-    """A convenience class to access cached time series datasets.
-
-    Note, that these *cached datasets* are statically included into *tslearn*
-    and are distinct from the ones in :class:`UCR_UEA_datasets`.
-
-    When using the Trace dataset, please cite [1]_.
-
-    See Also
-    --------
-    UCR_UEA_datasets : Provides more datasets and supports caching.
-
-    References
-    ----------
-    .. [1] A. Bagnall, J. Lines, W. Vickers and E. Keogh, The UEA & UCR Time
-       Series Classification Repository, www.timeseriesclassification.com
-    """
-    def __init__(self):
-        self.path = os.path.join(os.path.dirname(__file__), ".cached_datasets")
-
-    def list_datasets(self):
-        """List cached datasets.
-
-        Examples
-        --------
-        >>> _ = UCR_UEA_datasets().load_dataset("Trace")
-        >>> cached = UCR_UEA_datasets().list_cached_datasets()
-        >>> "Trace" in cached
-        True
-
-        Returns
-        -------
-        list of str:
-            A list of names of all cached (univariate and multivariate) dataset
-            namas.
-        """
-        return [fname[:fname.rfind(".")]
-                for fname in os.listdir(self.path)
-                if fname.endswith(".npz")]
-
-    def load_dataset(self, dataset_name):
-        """Load a cached dataset from its name.
-
-        Parameters
-        ----------
-        dataset_name : str
-            Name of the dataset. Should be in the list returned by
-            :meth:`~list_datasets`.
-
-        Returns
-        -------
-        numpy.ndarray of shape (n_ts_train, sz, d) or None
-            Training time series. None if unsuccessful.
-        numpy.ndarray of integers with shape (n_ts_train, ) or None
-            Training labels. None if unsuccessful.
-        numpy.ndarray of shape (n_ts_test, sz, d) or None
-            Test time series. None if unsuccessful.
-        numpy.ndarray of integers with shape (n_ts_test, ) or None
-            Test labels. None if unsuccessful.
-
-        Examples
-        --------
-        >>> data_loader = CachedDatasets()
-        >>> X_train, y_train, X_test, y_test = data_loader.load_dataset(
-        ...                                        "Trace")
-        >>> print(X_train.shape)
-        (100, 275, 1)
-        >>> print(y_train.shape)
-        (100,)
-
-        Raises
-        ------
-        IOError
-            If the dataset does not exist or cannot be read.
-        """
-        npzfile = numpy.load(os.path.join(self.path, dataset_name + ".npz"))
-        X_train = npzfile["X_train"]
-        X_test = npzfile["X_test"]
-        y_train = npzfile["y_train"]
-        y_test = npzfile["y_test"]
-        return X_train, y_train, X_test, y_test
