+++ tslearn/hdftools/__init__.py
@@ -1,0 +1,3 @@
+from .hdftools import save_dict, load_dict
+
+__all__ = ["save_dict", "load_dict"]
