+++ tslearn/preprocessing/__init__.py
@@ -1,0 +1,16 @@
+"""
+The :mod:`tslearn.preprocessing` module gathers time series scalers and 
+resamplers.
+"""
+
+from .preprocessing import (
+    TimeSeriesScalerMeanVariance,
+    TimeSeriesScalerMinMax,
+    TimeSeriesResampler
+)
+
+__all__ = [
+    "TimeSeriesResampler",
+    "TimeSeriesScalerMinMax",
+    "TimeSeriesScalerMeanVariance"
+]
